# openapi.cloudflare.yaml
openapi: 3.0.3
info:
  title: ExitRoute API
  description: A self-hosted informational API. It never authenticates to, operates, or stores credentials
    for a consumer account.
  version: 0.1.0
paths:
...
  /v1/webhook-subscriptions:
    post:
      tags:
      - webhooks
      summary: Create Subscription
      operationId: createWebhookSubscription
      security:
      - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookSubscriptionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionCreated'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags:
      - webhooks
      summary: List Subscriptions
      operationId: listWebhookSubscriptions
      security:
      - ApiKeyAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookSubscriptionView'
                title: Response Listwebhooksubscriptions
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/webhook-subscriptions/{subscription_id}:
    get:
      tags:
      - webhooks
      summary: Get Subscription
      operationId: getWebhookSubscription
      security:
      - ApiKeyAuth: []
      parameters:
      - name: subscription_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Subscription Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionView'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - webhooks
      summary: Delete Subscription
      operationId: deleteWebhookSubscription
      security:
      - ApiKeyAuth: []
      parameters:
      - name: subscription_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Subscription Id
      responses:
        '204':
          description: Successful Response
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/webhook-subscriptions/{subscription_id}/rotate-secret:
    post:
      tags:
      - webhooks
      summary: Rotate Subscription Secret
      operationId: rotateWebhookSecret
      security:
      - ApiKeyAuth: []
      parameters:
      - name: subscription_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Subscription Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSecretRotated'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  schemas:
...
    WebhookSecretRotated:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        signing_secret:
          type: string
          title: Signing Secret
      additionalProperties: false
      type: object
      required:
      - id
      - signing_secret
      title: WebhookSecretRotated
    WebhookSubscriptionCreate:
      properties:
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
        event_types:
          items:
            type: string
            enum:
            - published
            - superseded
            - stale
            - withdrawn
          type: array
          maxItems: 4
          minItems: 1
          title: Event Types
      additionalProperties: false
      type: object
      required:
      - url
      - event_types
      title: WebhookSubscriptionCreate
    WebhookSubscriptionCreated:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
        event_types:
          items:
            type: string
          type: array
          title: Event Types
        active:
          type: boolean
          title: Active
        created_at:
          type: string
          format: date-time
          title: Created At
        signing_secret:
          type: string
          title: Signing Secret
      additionalProperties: false
      type: object
      required:
      - id
      - url
      - event_types
      - active
      - created_at
      - signing_secret
      title: WebhookSubscriptionCreated
    WebhookSubscriptionView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
        event_types:
          items:
            type: string
          type: array
          title: Event Types
        active:
          type: boolean
          title: Active
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - url
      - event_types
      - active
      - created_at
      title: WebhookSubscriptionView
    Problem:
      additionalProperties: false
      properties:
        type:
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          title: Detail
          type: string
        code:
          title: Code
          type: string
        instance:
          default: null
          title: Instance
          type: string
          nullable: true
        request_id:
          default: null
          title: Request Id
          type: string
          nullable: true
        errors:
          default: null
          title: Errors
          items:
            additionalProperties: true
            type: object
          type: array
          nullable: true
      required:
      - type
      - title
      - status
      - detail
      - code
      title: Problem
      type: object
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: A scoped key created by an ExitRoute administrator.
      in: header
      name: X-API-Key
    BootstrapAdminAuth:
      type: http
      description: Bootstrap-only bearer token. Prefer a scoped admin API key after setup.
      scheme: bearer
  responses:
...
    Unauthorized:
      description: Authentication failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the required scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: The request conflicts with resource state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: The request violates a semantic or safety rule.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: The API-client quota was exceeded.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Seconds until a new quota window.
          schema:
            type: integer
            minimum: 1
    InternalError:
      description: The server could not complete the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...